wikis in yaml

+++ exo-addons-demo-extension-webapp/src/main/java/org/exoplatform/addons/populator/services/WikiService.java
@@ -1,5 +1,6 @@
 package org.exoplatform.addons.populator.services;
 
+import org.exoplatform.addons.populator.bean.WikiBean;
 import org.exoplatform.portal.config.model.PortalConfig;
 import org.exoplatform.wiki.mow.core.api.wiki.PageImpl;
 import org.exoplatform.wiki.resolver.TitleResolver;
@@ -8,6 +9,7 @@ import org.xwiki.rendering.syntax.Syntax;
 import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
 import javax.inject.Named;
+import java.util.List;
 import java.util.logging.Logger;
 
 @Named("wikiService")
@@ -23,31 +25,57 @@ public class WikiService {
     wikiService_ = wikiService;
   }
 
-  public void createUserWiki()
+  public void createUserWiki(List<WikiBean> wikis)
   {
-    createPage("Activity Stream Engagement", "activity-stream-engagement.txt");
-    createPage("How-To Guide", "how-to-guide.txt");
-    createPage("Leave Planning", "leave-planning.txt");
-    editPage("WikiHome", "wiki-home.txt");
+    for (WikiBean wiki:wikis)
+    {
+      createOrEditPage(wiki, wiki.getParent());
+    }
   }
 
 
-  private void createPage(String title, String filename)
+  private void createOrEditPage(WikiBean wikiBean, String parentTitle)
   {
-    createPage(title, filename, "WikiHome");
-  }
+    boolean forceNew = wikiBean.getNew();
+    String title= wikiBean.getTitle();
+    String filename = wikiBean.getFilename();
+    String parent = parentTitle;
+    try
+    {
+      if (forceNew && !title.equals("Wiki Home"))
+      {
+        if (wikiService_.isExisting(PortalConfig.PORTAL_TYPE, "intranet", TitleResolver.getId(title, false)))
+        {
+          wikiService_.deletePage(PortalConfig.PORTAL_TYPE, "intranet", TitleResolver.getId(title, false));
+        }
+      }
 
-  private void createPage(String title, String filename, String parent)
-  {
-    try {
+      PageImpl page;
       if (wikiService_.isExisting(PortalConfig.PORTAL_TYPE, "intranet", TitleResolver.getId(title, false)))
-        wikiService_.deletePage(PortalConfig.PORTAL_TYPE, "intranet", TitleResolver.getId(title, false));
+      {
+        page = (PageImpl) wikiService_.getPageById(PortalConfig.PORTAL_TYPE, "intranet", TitleResolver.getId(title, false));
+      }
+      else
+      {
+        page = (PageImpl) wikiService_.createPage(PortalConfig.PORTAL_TYPE, "intranet", title, TitleResolver.getId(parent, false));
+      }
 
-      PageImpl page = (PageImpl) wikiService_.createPage(PortalConfig.PORTAL_TYPE, "intranet", title, parent);
-      page.getContent().setText(Utils.getWikiPage(filename));
+      String content = "= "+title+" =";
+      if (filename!=null)
+        content = Utils.getWikiPage(filename);
+      page.getContent().setText(content);
       page.setSyntax(Syntax.XWIKI_2_0.toIdString());
       page.checkin();
       page.checkout();
+
+      if (wikiBean.getWikis()!=null && wikiBean.getWikis().size()>0)
+      {
+        for (WikiBean wiki:wikiBean.getWikis())
+        {
+          createOrEditPage(wiki, wikiBean.getTitle());
+        }
+      }
+
 
     } catch (Exception e) {
       e.printStackTrace();  //To change body of catch statement use File | Settings | File Templates.
@@ -55,19 +83,4 @@ public class WikiService {
 
   }
 
-  private void editPage(String title, String filename)
-  {
-    try {
-
-      PageImpl page = (PageImpl) wikiService_.getPageById(PortalConfig.PORTAL_TYPE, "intranet", TitleResolver.getId(title, false));
-      page.getContent().setText(Utils.getWikiPage(filename));
-      page.setSyntax(Syntax.XWIKI_2_0.toIdString());
-      page.checkin();
-      page.checkout();
-
-    } catch (Exception e) {
-      e.printStackTrace();  //To change body of catch statement use File | Settings | File Templates.
-    }
-
-  }
 }
